Refactor File System
