+++ NewHaven/src/GBMap.cpp
